requirements.txt
python-pptx==1.0.2
loguru==0.7.2
gradio==5.43.1

# LangChain Core Libraries
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.25

# LangChain Model Integrations
langchain-openai==0.3.31
langchain-ollama==0.3.7
langchain-deepseek-official==0.1.0.post1

python-docx==1.1.2
Pillow==9.1.0
torch==2.5.0
transformers==4.46.0
diffusers==0.35.1
invisible-watermark==0.2.0
safetensors==0.6.2
datasets==3.0.2
accelerate==1.0.1
librosa==0.10.2.post1
soundfile==0.12.1
ffmpeg==1.4
torchvision==0.20.0
sentencepiece==0.1.99
bitsandbytes==0.44.1
jupyterlab==4.3.0